app/utils/auth_checks.py
import dkim
import spf
import dnspython as dns
from cryptography.hazmat.primitives.asymmetric import rsa

def validate_auth(parsed: dict) -> dict:
    headers = parsed["headers"]
    # SPF
    spf_result = spf.check2(i=headers.get("Received").split()[1], s=headers.get("From"), h=headers.get("HELO"))
    # DKIM
    dkim_result = dkim.verify(headers.get("DKIM-Signature").encode())
    # DMARC: Stub with dns
    # Quantum-Resistant Check (Unique)
    if "DKIM-Signature" in headers:
        # Parse key from DNS
        selector = headers["DKIM-Signature"].split('s=')[1].split(';')[0]
        domain = headers["From"].split('@')[1]
        txt_record = dns.resolver.resolve(f"{selector}._domainkey.{domain}", "TXT")
        pub_key = txt_record[0].strings[0].decode()  # Simplified
        if "RSA" in pub_key and "bits<2048" :  # Check key strength
            quantum_flag = "Vulnerable"
        else:
            quantum_flag = "Resistant"
    return {"spf": spf_result, "dkim": dkim_result, "quantum": quantum_flag}